Data.Trie.Map.Internal

Description

This module exposes internal representation of TMap. TMap has one invariant condition:

Subtrees of an TMap should not be empty.

For example, consider following tree structure which is valid:

> fromList [("a",1), ("aa", 2), ("bc", 3)]
Root
  'a' -> 1
    'a' -> 2
  'b' ->
    'c' -> 3

Adding redundant node which represents empty map does not change what an TMap represents.

Root
  'a' -> 1
    'a' -> 2
  'b' ->
    'c' -> 3
  'd' ->
    'e' ->
      'f' ->

But such TMap should not exist because it confuses Eq and Ord instances and null function.

Types

newtype TMap c a

Mapping from [c] to a implemented as a trie. This type serves almost same purpose with Map [c] a, but can be looked up more efficiently.

Constructors

TMap

Fields

getNode :: Node c a (TMap c a)

data Node c a r

Constructors

Node !(Maybe a) !(Map c r)

foldTMap :: (Node c a r -> r) -> TMap c a -> r